+++ mitra_activitypub/src/builders/note.rs
@@ -4,7 +4,6 @@ use apx_sdk::{
     constants::{AP_MEDIA_TYPE, AP_PUBLIC},
     core::{
         multihash::encode_sha256_multihash,
-        url::http_uri::HttpUri,
     },
     deserialization::deserialize_string_array,
 };
@@ -169,19 +168,14 @@ pub struct Note {
 }
 
 pub fn build_note(
-    instance_uri: &HttpUri,
     instance_webfinger_hostname: &str,
     authority: &Authority,
     media_server: &MediaServer,
     post: &PostDetailed,
     with_context: bool,
 ) -> Note {
+    let server_uri = authority.expect_server_uri();
     let related_posts = post.expect_related_posts();
-    assert_eq!(
-        authority.expect_server_uri(),
-        instance_uri,
-        "authority should be anchored",
-    );
     let object_id = local_object_id_unified(authority, post.id);
     let mut object_type = NOTE;
     let actor_id = local_actor_id_unified(authority, post.author.id, &post.author.username);
@@ -267,7 +261,8 @@ pub fn build_note(
         tags.push(Tag::SimpleTag(tag));
     };
     for tag_name in &post.tags {
-        let tag_href = local_tag_collection(instance_uri.as_str(), tag_name);
+        // TODO: FEP-EF61: client should use server's URL template
+        let tag_href = local_tag_collection(server_uri.as_str(), tag_name);
         let tag = SimpleTag {
             tag_type: HASHTAG.to_string(),
             name: format!("#{}", tag_name),
@@ -301,7 +296,8 @@ pub fn build_note(
         .map(|linked| compatible_post_object_id(authority, linked));
 
     for emoji in &post.emojis {
-        let tag = build_emoji(instance_uri.as_str(), media_server, emoji);
+        // TODO: FEP-EF61: portable or anonymous emojis?
+        let tag = build_emoji(server_uri.as_str(), media_server, emoji);
         tags.push(Tag::EmojiTag(tag));
     };
 
@@ -338,9 +334,8 @@ pub fn build_note(
     };
     let conversation = post.expect_conversation();
     let maybe_context_id = if conversation.is_managed {
-        // TODO: FEP-EF61: use Authority
         let context_collection_id =
-            local_conversation_collection(instance_uri.as_str(), conversation.id);
+            local_conversation_collection(authority, conversation.id);
         Some(context_collection_id)
     } else {
         conversation.object_id.clone()
@@ -428,6 +423,7 @@ pub async fn get_note_recipients(
 
 #[cfg(test)]
 mod tests {
+    use apx_core::url::http_uri::HttpUri;
     use serde_json::json;
     use uuid::uuid;
     use mitra_models::{
@@ -471,7 +467,6 @@ mod tests {
         let authority = Authority::server(&instance_uri);
         let media_server = MediaServer::for_test(INSTANCE_URI);
         let note = build_note(
-            &instance_uri,
             INSTANCE_HOSTNAME,
             &authority,
             &media_server,
@@ -550,7 +545,6 @@ mod tests {
         let authority = Authority::server(&instance_uri);
         let media_server = MediaServer::for_test(INSTANCE_URI);
         let question = build_note(
-            &instance_uri,
             INSTANCE_HOSTNAME,
             &authority,
             &media_server,
@@ -609,7 +603,6 @@ mod tests {
         let authority = Authority::server(&instance_uri);
         let media_server = MediaServer::for_test(INSTANCE_URI);
         let note = build_note(
-            &instance_uri,
             INSTANCE_HOSTNAME,
             &authority,
             &media_server,
@@ -640,7 +633,6 @@ mod tests {
         let authority = Authority::server(&instance_uri);
         let media_server = MediaServer::for_test(INSTANCE_URI);
         let note = build_note(
-            &instance_uri,
             INSTANCE_HOSTNAME,
             &authority,
             &media_server,
@@ -672,7 +664,6 @@ mod tests {
         let authority = Authority::server(&instance_uri);
         let media_server = MediaServer::for_test(INSTANCE_URI);
         let note = build_note(
-            &instance_uri,
             INSTANCE_HOSTNAME,
             &authority,
             &media_server,
@@ -699,7 +690,6 @@ mod tests {
         let authority = Authority::server(&instance_uri);
         let media_server = MediaServer::for_test(INSTANCE_URI);
         let note = build_note(
-            &instance_uri,
             INSTANCE_HOSTNAME,
             &authority,
             &media_server,
@@ -753,7 +743,6 @@ mod tests {
         let authority = Authority::server(&instance_uri);
         let media_server = MediaServer::for_test(INSTANCE_URI);
         let note = build_note(
-            &instance_uri,
             INSTANCE_HOSTNAME,
             &authority,
             &media_server,
@@ -825,7 +814,6 @@ mod tests {
         let authority = Authority::server(&instance_uri);
         let media_server = MediaServer::for_test(INSTANCE_URI);
         let note = build_note(
-            &instance_uri,
             INSTANCE_HOSTNAME,
             &authority,
             &media_server,
@@ -903,7 +891,6 @@ mod tests {
         );
         let media_server = MediaServer::for_test(INSTANCE_URI);
         let note = build_note(
-            &instance_uri,
             INSTANCE_HOSTNAME,
             &authority,
             &media_server,
@@ -928,7 +915,7 @@ mod tests {
             "attributedTo": "https://server.example/.well-known/apgateway/did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actors/46d160ae-af12-484d-9f44-419f00fc1b31",
             "content": "",
             "sensitive": false,
-            "context": "https://server.example/collections/conversations/837ffc24-dab2-414b-a9b8-fe47d0a463f2",
+            "context": "https://server.example/.well-known/apgateway/did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/collections/conversations/837ffc24-dab2-414b-a9b8-fe47d0a463f2",
             "replies": "https://server.example/.well-known/apgateway/did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/objects/11fa64ff-b5a3-47bf-b23d-22b360581c3f/replies",
             "tag": [{
                 "type": "Mention",