+++ mitra_api/src/mastodon_api/statuses/views.rs
@@ -1230,7 +1230,6 @@ async fn make_permanent(
     let authority = Authority::server(instance.uri());
     let media_server = MediaServer::new(&config);
     let note = build_note(
-        instance.uri(),
         &instance.webfinger_hostname(),
         &authority,
         &media_server,